Added tsconfig.json and fixed all errors.
